+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/_core2/__init__.py
@@ -13,3 +13,6 @@ __all__: tuple[str, ...] = (
     "WorkflowContext",
     "WorkflowRunnerManager",
 )
+
+
+# TODO: there as still issues with the name of this module
