+++ services/director-v2/tests/unit/dynamic_scheduler/conftest.py
@@ -2,10 +2,10 @@
 
 import pytest
 from pytest import FixtureRequest
-from simcore_service_director_v2.modules.dynamic_sidecar.scheduler._v2._context_base import (
+from simcore_service_director_v2.modules.dynamic_sidecar.scheduler._core2._context_base import (
     ContextInterface,
 )
-from simcore_service_director_v2.modules.dynamic_sidecar.scheduler._v2._context_in_memory import (
+from simcore_service_director_v2.modules.dynamic_sidecar.scheduler._core2._context_in_memory import (
     InMemoryContext,
 )
