+++ services/director-v2/tests/unit/dynamic_scheduler/test_modules_dynamic_sidecar_scheduler__core2__context_in_memory.py
@@ -4,7 +4,7 @@ from typing import Any
 
 import pytest
 from pytest import FixtureRequest
-from simcore_service_director_v2.modules.dynamic_sidecar.scheduler._v2._context_in_memory import (
+from simcore_service_director_v2.modules.dynamic_sidecar.scheduler._core2._context_in_memory import (
     InMemoryContext,
 )
