+++ services/director-v2/tests/unit/dynamic_scheduler/test_modules_dynamic_sidecar_scheduler__core2_workflow.py
@@ -1,12 +1,12 @@
 import pytest
-from simcore_service_director_v2.modules.dynamic_sidecar.scheduler._v2._action import (
+from simcore_service_director_v2.modules.dynamic_sidecar.scheduler._core2._action import (
     Action,
 )
-from simcore_service_director_v2.modules.dynamic_sidecar.scheduler._v2._errors import (
+from simcore_service_director_v2.modules.dynamic_sidecar.scheduler._core2._errors import (
     NextActionNotInWorkflowException,
     OnErrorActionNotInWorkflowException,
 )
-from simcore_service_director_v2.modules.dynamic_sidecar.scheduler._v2._workflow import (
+from simcore_service_director_v2.modules.dynamic_sidecar.scheduler._core2._workflow import (
     Workflow,
 )
